IW-3184 | Reduce number of parsoid workers and increase replicas
Parsoid uses the service-runner library to start up a number of workers
to serve traffic. By default, it starts one worker per available CPU core.
In a container context, this is not desired, as the library reads the
available CPU cores on the host, rather than the actual CPU resources
available to the container after resource limits are applied.
Thus, we end up spawning too many processes, exhausting the memory limit
of the container.

As a fix, use a fixed low number of workers (4) per container,
and increase the number of deployment replicas to improve resiliency.