CATS-270 | Account for template content before portable infoboxes
Parsoid represents encapsulated content such as template transclusions as a
forest of DOM trees formed by adjacent DOM nodes, where the first node holds the
type and metadata of the encapsulation and subsequent nodes refer to it via an
RDFa about ID. For portable infobox templates, the first node is generally the
infobox itself, being the first or only piece of content in the template, but
there may be other content in the template, such as a category link, preceding
the infobox. This confuses our portable infobox handler, since it assumes that
infoboxes in template transclusions are always the first node inside the
transclusion, so templates that do not fit this expectation are ignored during
rendering.

As a fix, look for the template transclusion start node when processing
potential infobox transclusions during DOM post-processing, instead of assuming
it's always going to be the same node as the infobox itself. Since this means
that the template parameters may no longer be co-located on the infobox node,
store a 2-tuple of [transclusionStartNode,infoboxNode] for each processed
transclusion, so that both the template parameters and the destination node
remain available for subsequent rendering and substitution.