CATS-194 | Don't manipulate transcluded main page column DSR offsets
Our current main page column tag handler crashes when trying to handle templated
main pages, since it tries to use the DSR offsets of the end column tag to
correct the DSR offsets for the start tag, but the end column tag will have no
offsets set in this scenario. Instead, make sure not to mess with DSRs if the
main page tags came from a transclusion - their DSR offsets will be correct in
the first place, since they will refer to the offsets of the template
transclusion in the main page wikitext, rather than the offsets of the tags in
the template wikitext, thereby requiring no adjustment. Add two parser tests for
templated main pages as well.

Ultimately, none of this code is idiomatic Parsoid; all the DSR etc. voodoo is
only needed to work around the fact that we use a separate "end tag" (that is,
<mainpage-endcolumn />) to close off main page columns, rather than going the
idiomatic way and having <mainpage-leftcolumn>...</mainpage-leftcolumn> tags
that can have wikitext content. In the latter case, Parsoid handling could be
achieved with either a trivial native extension or by the builtin fallback to
PHP parser handling.