CATS-1086 | Refactor portable infobox parsing for robustness
Portable infoboxes need access to the current parser frame so that they can use
template parameters for rendering infobox data. To support this, we implement
portable infobox parsing in Parsoid using an HTML post-processor that runs after
Parsoid has assembled a mostly-ready DOM for the article being parsed. This
processor looks for portable infoboxes within the DOM, finds associated template
transclusion parameters, and calls out to the PHP parser to perform the actual
rendering.

The current logic proactively looks for <aside> node (the root element of
portable infoboxes), then checks if the node itself or one of its siblings is an
associated template transclusion. This fails to account for the case where the
infobox may be "buried" as a child node of a different node within the template
transclusion - in this case, the associated template transclusion may be an
ancestor of the infobox node, or a sibling of one of its ancestors.

As such, refactor portable infobox parsing to look for template transclusions
first and then try to find associated portable infoboxes (which may be the
transclusion node itself, a sibling of the transclusion node, a child of the
transclusion node, or a child of one of the transclusion node's siblings),
rather than the other way around. Incidentally, this should also correctly
handle the rare (but theoretically possible) case where a template transclusion
has multiple associated portable infoboxes; the current logic would always use
the rendered HTML of the first infobox if such a situation occurred.

Also add some basic parser tests for portable infobox parsing.