Switch Parsoid to new ingress
Make Parsoid use external-lb-a instead of traefik. This is only relevant
in dev, since production UCP calls Parsoid internally via k8s service.

Also ditch path prefix stripping in favor of a custom subdomain. Path
prefix strip was never a good fit for Parsoid to begin with, as it broke
the development tools and forms due to incorrect paths on the server
side.

MW development configuration will be updated next.